% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_methods.R
\name{compare_methods}
\alias{compare_methods}
\title{GLM-based GWAS Analysis}
\usage{
compare_methods(n = 100, m = 1000, replicates = 30)
}
\arguments{
\item{y}{Phenotype data (n x 1).}

\item{X}{Genotype data (n x m).}

\item{C}{Covariate data (n x t).}
}
\value{
A numeric vector of p-values (1 x m).
}
\description{
Performs genome-wide association studies (GWAS) using a generalized linear model (GLM).
}
