% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_GWAS.R
\name{GLM_GWAS}
\alias{GLM_GWAS}
\title{GLM-based GWAS Analysis}
\usage{
GLM_GWAS(y, X, C)
}
\arguments{
\item{y}{Phenotype data (n x 1).}

\item{X}{Genotype data (n x m).}

\item{C}{Covariate data (n x t).}
}
\value{
A numeric vector of p-values (1 x m).
}
\description{
Performs genome-wide association studies (GWAS) using a generalized linear model (GLM).
}
