% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_GWAS.R
\name{PCA_GWAS}
\alias{PCA_GWAS}
\title{GLM-based GWAS Analysis}
\usage{
PCA_GWAS(X, C, num_PCs = 5)
}
\arguments{
\item{X}{Genotype data (n x m).}

\item{C}{Covariate data (n x t).}

\item{y}{Phenotype data (n x 1).}
}
\value{
A numeric vector of p-values (1 x m).
}
\description{
Performs genome-wide association studies (GWAS) using a generalized linear model (GLM).
}
