% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{GLM-based GWAS Analysis}
\usage{
simulate_data(n, m, causal_snp = 10)
}
\arguments{
\item{y}{Phenotype data (n x 1).}

\item{X}{Genotype data (n x m).}

\item{C}{Covariate data (n x t).}
}
\value{
A numeric vector of p-values (1 x m).
}
\description{
Performs genome-wide association studies (GWAS) using a generalized linear model (GLM).
}
